refactor: split filesystem helpers from process IO
